rename package utils to util
